Completely Revamped Project
Response: White Illiteracy Rate
Predictors: not good, Total Number of White Illiterate Persons is directly related to the White Illiteracy Rate.
The only predictor that might be worthwhile was a binary predictor I created `Confederate` which is 1 if the county fought against the United States (The Union) as part of the Confederacy.
Bunch of potential predictors were weakly correlated, positively or negatively.